feat: implement getSourceWithoutMargins
